refactor: move errors useful to clients (for retries etc) outside of internal so they can be imported
